# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Proyectos",
        "version": "1.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Lista todas las tareas",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Listado de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Corregir bug"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pendiente"
                                            },
                                            "priority": {
                                                "type": "string",
                                                "example": "alta"
                                            },
                                            "estimated_hours": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5.5
                                            },
                                            "actual_hours": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 3
                                            },
                                            "due_date": {
                                                "type": "string",
                                                "example": "2025-12-31"
                                            },
                                            "project_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "assigned_user_id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una tarea",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status",
                                    "priority",
                                    "estimated_hours",
                                    "due_date",
                                    "project_id",
                                    "assigned_user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Corregir bug"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del bug..."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "alta"
                                    },
                                    "estimated_hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "example": "2025-12-31"
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assigned_user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Elimina una Tarea",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarea eliminada"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        },
...
        "/api/teams/{id}": {
            "get": {
                "tags": [
                    "Equipos"
                ],
                "summary": "Obtener información de un equipo",
                "operationId": "ab05003decb5cbc16154382836f51e3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del equipo"
                    },
                    "404": {
                        "description": "Equipo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Equipos"
                ],
                "summary": "Elimina un equipo",
                "operationId": "9b992594932c3e54aa58b58c3c1e933c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Equipo eliminado"
                    },
                    "404": {
                        "description": "Equipo no encontrado"
                    }
                }
            }
        },
...
    },
...
}